% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFeaturesRanking.R
\name{ensemble.features.ranking}
\alias{ensemble.features.ranking}
\title{Enhanced Ensemble Features Ranking}
\usage{
ensemble.features.ranking(dataset, ...)
}
\arguments{
\item{dataset}{dataset with numerial features and binary categorical class, named 'class'}

\item{nRows}{number of rows of each dataset partition, 10 default value}

\item{nTries}{number of runs, ncol(dataset) / 2, default value}

\item{cutOff}{k = n (selects n features); k=0 (selects all features); k=-1 (automatic selecion) default}

\item{methods}{list of filter functions (ensemble) for feature evaluation, default list(gain.ratio,symmetrical.uncertainty,chi.squared, random.forest.importance)}
}
\value{
features names list inversely ordered by relevance
}
\description{
Takes a dataset and outputs a features name list inversely ordered by relevance
}
\examples{
\dontrun{
}
data(allDataProbe)
library(FSelector)
set.seed(134)
methods <- list(gain.ratio,chi.squared)
features <- ensemble.features.ranking(allDataProbe, nRows=nrow(allDataProbe)/2, nTries=20, cutOff= -1, methods=methods)
print(features)

}
