% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRowSamples.R
\name{getNRandomsRowsSubsets}
\alias{getNRandomsRowsSubsets}
\title{Get N random Saples}
\usage{
getNRandomsRowsSubsets(nRows, size, iterations)
}
\arguments{
\item{nRows}{dataset number of rows}

\item{size}{subset number of rows}

\item{interactions}{number of subsets (N)}
}
\value{
N subsets of row indexs
}
\description{
Get N random row indexes subsets
}
